Allow including grd files from other grd files with <part file="sub.grp"/>
The top-level element of included files must be <grit-part>. Subelements of <grit-part> behave as though they were spliced in at the location of the <part> node in the including file. The extension ".grp" is suggested for included files.

Included files have to be in the same directory as the parent, to avoid complications related to relative paths.

<part> nodes are allowed in the same places as <if> nodes. This may be too restrictive.

BUG=142381
TEST=new unit tests

Review URL: https://codereview.chromium.org/11046013